Rename: UserStudyClickListener ui 패키지로 이동
