+++ app/src/main/java/com/sesac/developer_study_platform/ui/UserStudyClickListener.kt
@@ -1,4 +1,4 @@
-package com.sesac.developer_study_platform
+package com.sesac.developer_study_platform.ui
 
 import com.sesac.developer_study_platform.data.UserStudy
