+++ app/src/main/java/com/sesac/developer_study_platform/ui/home/StudyAdapter.kt
@@ -7,7 +7,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.sesac.developer_study_platform.R
-import com.sesac.developer_study_platform.UserStudyClickListener
+import com.sesac.developer_study_platform.ui.UserStudyClickListener
 import com.sesac.developer_study_platform.data.UserStudy
 import com.sesac.developer_study_platform.databinding.ItemStudyBinding
